vuln_analyzer-GUI/requirements.txt
flask
python-dotenv
openai
anthropic
gunicorn
werkzeug
requests
colorama
flask-wtf
flask-cors
